Database work up to INV
Got started in on the database service for populating the cards table. This service behaves a little differently in order to adhere to proper database connection principles.

Some fancy footwork required for a few cases that came up before the invasion of split-cards. Foreign cards and special gold-bordered/backed cards are skipped for purity's sake, and skipping in general requires a couple of approaches for how to go about counting and recording info into the manifest. Additional escaping was required as was expanding name-based database columns for Unglued's weirdness. Image work needed a bit of tuning to account for skipping entire sets. Lastly, the Manifest service got an upgrade to support caring about the until-now-unused diff column.